Split projects and plots
